Apply sauce creds by overwriting .travis.yml
