+++ build-system/sauce-credentials.json
@@ -11,7 +11,7 @@
     },
     "access_key_encrypted": {
       "_description": "Your encrypted Sauce Labs access key",
-      "_instructions": "Copy your access key from https://saucelabs.com/beta/users/<username> and run `travis encrypt SAUCE_ACCESS_KEY=<your key>`"
+      "_instructions": "Copy your access key from https://saucelabs.com/beta/users/<username> and run `travis encrypt -r ampproject/amphtml SAUCE_ACCESS_KEY=\"<your key>\"`"
     }
   },
   "credentials": {
@@ -21,7 +21,7 @@
     },
     "rsimha@ampproject.org": {
       "username": "rsimha",
-      "access_key_encrypted": "qm6uXnJGwX+W/lXBaXZNfde9jweKag/ZG82BCI16tUvuDpxVNPQGvMCF5xhc81G9SXl5zGqOaUiXe2ApZ+1rhkPmeGoI8+z+M1N0Rd0BuP9PfUPC3gWzKR6vnckPa68R3AycLkzzo2AdeDyqxcrpheqSXIoIshejvJ6CQuJw3gbbKp/Adnp5mpojH296ubdrGymRFDpk6TdDY1BRF0Mk7+1vQH86e/DNLnwt+FexEuRZ4Fde7PcQLVRRnTIOVTQ4dLkp1cBzzExAWl44lrIuo6tdFcD+8wayDxVnb/HPWf7CWb9EQTk9+bN5yZhnhvfeHyJIwuRItq4UgTngVCwN4Pc44wpkWHeee2KukdS2DCtMzCjHQFvpGFuCbEx8ulAS9psjYJQSVfXvpLOpQmdi8mRklrGN0QaGMPlTD2UwVeoD5CNBwQXlhgImb6PXB/sFBQyXau7YUNV1FcJnY7UIIvAPgcoLjERqo3mW8U5gPC5Luvb8GRFSlC95jUtrDCSofD/bIJ2pxWazbhPlfiusT2ENohTxL90CbXr236hFcM/H+aWWqF3dKAlMJ44JDX0aBZlG2GF5yQ1W8vn3cJXvv4Qu42Me8pkPnw26Z1xK0Chap1lfjNuxUFnSwA5QG7RIRPhYru0suOPEjmMg/YVLwqdv8Vvmr/fRB38qLdoClaQ="
+      "access_key_encrypted": "HnJQHcQIegWPAAm5K2KY7Y/4xZoMS2nodlUE+GCMxBhG7+iIjqfaas1gz+bfgzIqVqgkSeLUy8LbHuJ9llA0rwi4/NxQUxp6djAWAfnYnO/PUkmejAEmwPL1aUk2+ZGUkglfrbXfNij/z0T2AW/YoJOHJfa/DQecGl6COqjQohYtQ64D5wvwOAiNZ/zLHlQS1q5IETBGw18FkF/4nLuEy3mCCZJ6Aa/1LQ+43/FsxRYD448lD1FVjphDofRbfUNB4K6urwfvCYs3g5rl+2BA89jhsHSRA9uQKkBsjkh8RinN098j5r9+wBGLe8GOAY0V36COHKswEbs15ISRI7wR5jIqxXfP7wXhdUDuBdCrmZjpoNihclNl8VST12FiOvfBqQqQDypKEtt+fZlK6KeSZ4rZqMRPdK8Vq17oRBkQ1y9j1tg7I4Z1aTeGH2/hPUzVteZJtqYta3IKQjJSwJeVq3gnH2Epg6V2twDutoUZwsn15dAjh/17oZKwpcATnSW1Oni/H00NqG8qBaIcV6GOeKv/YpbWWdGCKKs4RLx6L2IRvs9tMt993cl12BwBSl0Y9ROCRl8eCPlKBipBnEUVYVuN6QhWxQlBuUnFqwXGXF32ZYv4UNTNkOkE4i3xbQqLCnNCKSxEucAdANx1ajxmAOLRa3Bj+HbWTrWhH7HKQWo="
     }
   }
 }
